Interruption handlers for Download
Changed: Utils [module]
Purpose: Handling interruption of download.
Detail: If something happens while download is going, python will automatically
save the file as if it was fully completed, which may lead to corruption. To
protect us from that instead of saving the file as it is, we're saving it to
temporary file which will then be renamed as the target file.